+++ mathics/builtin/drawing/__init__.py
@@ -1,0 +1,3 @@
+"""
+Graphics, Drawing, and Images
+"""