+++ mathics/builtin/numeric.py
@@ -37,6 +37,7 @@ from mathics.core.expression import (
     Complex,
     Expression,
     Integer,
+    Integer0,
     MachineReal,
     Number,
     Rational,
@@ -1143,13 +1144,13 @@ class Rationalize(Builtin):
 def chop(expr, delta=10.0 ** (-10.0)):
     if isinstance(expr, Real):
         if -delta < expr.get_float_value() < delta:
-            return Integer(0)
+            return Integer0
     elif isinstance(expr, Complex) and expr.is_inexact():
         real, imag = expr.real, expr.imag
         if -delta < real.get_float_value() < delta:
-            real = Integer(0)
+            real = Integer0
         if -delta < imag.get_float_value() < delta:
-            imag = Integer(0)
+            imag = Integer0
         return Complex(real, imag)
     elif isinstance(expr, Expression):
         return Expression(chop(expr.head), *[chop(leaf) for leaf in expr.leaves])